fix: Выносит Compound Components в отдельные файлы, оборачивает memo(Comment), уносить стили для селектов ближе к компоненту, выносит логику сортировки в хендлер (сортировка по длине)
