+++ src/features/PostLengthSorted/ui/PostLengthSorted.module.css
@@ -1,13 +1,3 @@
-.input {
-  transition: var(--trans-default);
-
-  &:hover,
-  &:focus {
-    border-color: #535bf2;
-    opacity: 1;
-  }
-}
-
 .selectControl {
   padding: 3px 5px;
   border: none;
